# replit-app/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HuskQueryInput {
  /** @minLength 1 */
  query: string;
}

export type HuskQueryResultVerdict = typeof HuskQueryResultVerdict[keyof typeof HuskQueryResultVerdict];


export const HuskQueryResultVerdict = {
  answered: 'answered',
  silenced: 'silenced',
} as const;

/**
 * Non-null when the model tried to answer and the server-side gate vetoed it ("distance" = too far, "overlap" = no meaningful lexical overlap with the shard). Null when the model abstained on its own, or when the query was answered.
 * @nullable
 */
export type HuskQueryResultGateVeto = typeof HuskQueryResultGateVeto[keyof typeof HuskQueryResultGateVeto] | null;


export const HuskQueryResultGateVeto = {
  distance: 'distance',
  overlap: 'overlap',
} as const;

export interface HuskQueryResult {
  query: string;
  silenced: boolean;
  verdict: HuskQueryResultVerdict;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  citations?: string | null;
  /** @nullable */
  concept?: string | null;
  /**
     * Best normalized Hamming distance measured by the gate
     * @nullable
     */
  distance?: number | null;
  threshold?: number;
  /**
     * Non-null when the model tried to answer and the server-side gate vetoed it ("distance" = too far, "overlap" = no meaningful lexical overlap with the shard). Null when the model abstained on its own, or when the query was answered.
     * @nullable
     */
  gateVeto?: HuskQueryResultGateVeto;
  /**
     * Content tokens shared between the query side and the candidate shard
     * @nullable
     */
  sharedTokens?: number | null;
  fingerprint?: string;
  /** Number of agent loop turns taken */
  turns?: number;
}

export interface HuskShard {
  id: number;
  concept: string;
  source: string;
  text: string;
  fingerprint: string;
  createdAt?: string;
}

export interface HuskDecision {
  id: number;
  query: string;
  fingerprint: string;
  /** @nullable */
  bestDistance?: string | null;
  threshold: string;
  verdict: string;
  /** @nullable */
  concept?: string | null;
  /** @nullable */
  toolCalls?: string | null;
  createdAt?: string;
}

export interface HealthStatus {
  status: string;
}