# replit-app/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /husk/query:
    post:
      operationId: huskQuery
      tags: [husk]
      summary: Query the Husk agent harness
      description: Multi-turn tool-calling loop over the SimHash-128 shard vault. Returns verified shard text or deliberate silence with the measured Hamming distance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HuskQueryInput"
      responses:
        "200":
          description: Shard result or silence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HuskQueryResult"
...
components:
  schemas:
    HuskQueryInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
    HuskQueryResult:
      type: object
      required: [query, silenced, verdict]
      properties:
        query:
          type: string
        silenced:
          type: boolean
        verdict:
          type: string
          enum: [answered, silenced]
        text:
          type: ["string", "null"]
        citations:
          type: ["string", "null"]
        concept:
          type: ["string", "null"]
        distance:
          type: ["number", "null"]
          description: Best normalized Hamming distance measured by the gate
        threshold:
          type: number
        gateVeto:
          type: ["string", "null"]
          enum: [distance, overlap, null]
          description: >-
            Non-null when the model tried to answer and the server-side gate
            vetoed it ("distance" = too far, "overlap" = no meaningful
            lexical overlap with the shard). Null when the model abstained on
            its own, or when the query was answered.
        sharedTokens:
          type: ["integer", "null"]
          description: Content tokens shared between the query side and the candidate shard
        fingerprint:
          type: string
        turns:
          type: integer
          description: Number of agent loop turns taken
...