# replit-app/lib/api-zod/src/generated/types/huskQueryInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HuskQueryInput {
  /** @minLength 1 */
  query: string;
}