# replit-app/lib/api-zod/src/generated/types/huskQueryResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HuskQueryResultGateVeto } from './huskQueryResultGateVeto';
import type { HuskQueryResultVerdict } from './huskQueryResultVerdict';

export interface HuskQueryResult {
  query: string;
  silenced: boolean;
  verdict: HuskQueryResultVerdict;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  citations?: string | null;
  /** @nullable */
  concept?: string | null;
  /**
     * Best normalized Hamming distance measured by the gate
     * @nullable
     */
  distance?: number | null;
  threshold?: number;
  /**
     * Non-null when the model tried to answer and the server-side gate vetoed it ("distance" = too far, "overlap" = no meaningful lexical overlap with the shard). Null when the model abstained on its own, or when the query was answered.
     * @nullable
     */
  gateVeto?: HuskQueryResultGateVeto;
  /**
     * Content tokens shared between the query side and the candidate shard
     * @nullable
     */
  sharedTokens?: number | null;
  fingerprint?: string;
  /** Number of agent loop turns taken */
  turns?: number;
}